+++ packages/liveblocks-react/src/factory.tsx
@@ -19,6 +19,16 @@ import * as React from "react";
 import { useClient as _useClient } from "./client";
 import useRerender from "./useRerender";
 
+/**
+ * "Freezes" a given value, so that it will return the same value/instance on
+ * each subsequent render. This can be used to freeze "initial" values for
+ * custom hooks, much like how `useState(initialState)` or
+ * `useRef(initialValue)` works.
+ */
+function useInitial<T>(value: T): T {
+  return React.useRef(value).current;
+}
+
 export type RoomProviderProps<
   TPresence extends JsonObject,
   TStorage extends LsonObject
@@ -232,13 +242,13 @@ type RoomContext<
    * @example
    * const animals = useList("animals");  // e.g. [] or ["🦁", "🐍", "🦍"]
    */
-  deprecated_useList<TValue extends Lson>(key: string): LiveList<TValue> | null;
+  useList_deprecated<TValue extends Lson>(key: string): LiveList<TValue> | null;
 
   /**
    * @deprecated We no longer recommend initializing the
    * items from the useList() hook. For details, see https://bit.ly/3Niy5aP.
    */
-  deprecated_useList<TValue extends Lson>(
+  useList_deprecated<TValue extends Lson>(
     key: string,
     items: TValue[]
   ): LiveList<TValue> | null;
@@ -253,7 +263,7 @@ type RoomContext<
    * @example
    * const shapesById = useMap("shapes");
    */
-  deprecated_useMap<TKey extends string, TValue extends Lson>(
+  useMap_deprecated<TKey extends string, TValue extends Lson>(
     key: string
   ): LiveMap<TKey, TValue> | null;
 
@@ -261,7 +271,7 @@ type RoomContext<
    * @deprecated We no longer recommend initializing the
    * entries from the useMap() hook. For details, see https://bit.ly/3Niy5aP.
    */
-  deprecated_useMap<TKey extends string, TValue extends Lson>(
+  useMap_deprecated<TKey extends string, TValue extends Lson>(
     key: string,
     entries: readonly (readonly [TKey, TValue])[] | null
   ): LiveMap<TKey, TValue> | null;
@@ -276,7 +286,7 @@ type RoomContext<
    * @example
    * const object = useObject("obj");
    */
-  deprecated_useObject<TData extends LsonObject>(
+  useObject_deprecated<TData extends LsonObject>(
     key: string
   ): LiveObject<TData> | null;
 
@@ -284,7 +294,7 @@ type RoomContext<
    * @deprecated We no longer recommend initializing the fields from the
    * useObject() hook. For details, see https://bit.ly/3Niy5aP.
    */
-  deprecated_useObject<TData extends LsonObject>(
+  useObject_deprecated<TData extends LsonObject>(
     key: string,
     initialData: TData
   ): LiveObject<TData> | null;
@@ -346,6 +356,15 @@ export function createRoomContext<
 
     const _client = useClient();
 
+    // Note: We'll hold on to the initial value given here, and ignore any
+    // changes to this argument in subsequent renders
+    const frozen = useInitial({
+      initialPresence,
+      initialStorage,
+      defaultPresence, // Will get removed in 0.18
+      defaultStorageRoot, // Will get removed in 0.18
+    });
+
     const [room, setRoom] = React.useState<
       Room<TPresence, TStorage, TUserMeta, TRoomEvent>
     >(() =>
@@ -361,10 +380,10 @@ export function createRoomContext<
     React.useEffect(() => {
       setRoom(
         _client.enter(roomId, {
-          initialPresence,
-          initialStorage,
-          defaultPresence, // Will get removed in 0.18
-          defaultStorageRoot, // Will get removed in 0.18
+          initialPresence: frozen.initialPresence,
+          initialStorage: frozen.initialStorage,
+          defaultPresence: frozen.defaultPresence, // Will get removed in 0.18
+          defaultStorageRoot: frozen.defaultStorageRoot, // Will get removed in 0.18
           DO_NOT_USE_withoutConnecting: typeof window === "undefined",
         } as RoomInitializers<TPresence, TStorage>)
       );
@@ -372,7 +391,7 @@ export function createRoomContext<
       return () => {
         _client.leave(roomId);
       };
-    }, [_client, roomId]);
+    }, [_client, roomId, frozen]);
 
     return <RoomCtx.Provider value={room}>{props.children}</RoomCtx.Provider>;
   }
@@ -398,7 +417,7 @@ export function createRoomContext<
       return () => {
         unsubscribe();
       };
-    }, [room]);
+    }, [room, rerender]);
 
     const setPresence = React.useCallback(
       (overrides: Partial<TPresence>, options?: { addToHistory: boolean }) =>
@@ -432,7 +451,7 @@ export function createRoomContext<
       return () => {
         unsubscribe();
       };
-    }, [room]);
+    }, [room, rerender]);
 
     return room.getOthers();
   }
@@ -509,7 +528,7 @@ export function createRoomContext<
         unsubscribePresence();
         unsubscribeConnection();
       };
-    }, [room]);
+    }, [room, rerender]);
 
     return room.getSelf();
   }
@@ -538,14 +557,14 @@ export function createRoomContext<
     return [root];
   }
 
-  function deprecated_useMap<TKey extends string, TValue extends Lson>(
+  function useMap_deprecated<TKey extends string, TValue extends Lson>(
     key: string
   ): LiveMap<TKey, TValue> | null;
-  function deprecated_useMap<TKey extends string, TValue extends Lson>(
+  function useMap_deprecated<TKey extends string, TValue extends Lson>(
     key: string,
     entries: readonly (readonly [TKey, TValue])[] | null
   ): LiveMap<TKey, TValue> | null;
-  function deprecated_useMap<TKey extends string, TValue extends Lson>(
+  function useMap_deprecated<TKey extends string, TValue extends Lson>(
     key: string,
     entries?: readonly (readonly [TKey, TValue])[] | null | undefined
   ): LiveMap<TKey, TValue> | null {
@@ -597,14 +616,14 @@ Please see https://bit.ly/3Niy5aP for details.`
     }
   }
 
-  function deprecated_useList<TValue extends Lson>(
+  function useList_deprecated<TValue extends Lson>(
     key: string
   ): LiveList<TValue> | null;
-  function deprecated_useList<TValue extends Lson>(
+  function useList_deprecated<TValue extends Lson>(
     key: string,
     items: TValue[]
   ): LiveList<TValue> | null;
-  function deprecated_useList<TValue extends Lson>(
+  function useList_deprecated<TValue extends Lson>(
     key: string,
     items?: TValue[] | undefined
   ): LiveList<TValue> | null {
@@ -658,14 +677,14 @@ Please see https://bit.ly/3Niy5aP for details.`
     }
   }
 
-  function deprecated_useObject<TData extends LsonObject>(
+  function useObject_deprecated<TData extends LsonObject>(
     key: string
   ): LiveObject<TData> | null;
-  function deprecated_useObject<TData extends LsonObject>(
+  function useObject_deprecated<TData extends LsonObject>(
     key: string,
     initialData: TData
   ): LiveObject<TData> | null;
-  function deprecated_useObject<TData extends LsonObject>(
+  function useObject_deprecated<TData extends LsonObject>(
     key: string,
     initialData?: TData
   ): LiveObject<TData> | null {
@@ -722,19 +741,19 @@ Please see https://bit.ly/3Niy5aP for details.`
   function useList<TKey extends Extract<keyof TStorage, string>>(
     key: TKey
   ): TStorage[TKey] | null {
-    return deprecated_useList(key) as unknown as TStorage[TKey];
+    return useList_deprecated(key) as unknown as TStorage[TKey];
   }
 
   function useMap<TKey extends Extract<keyof TStorage, string>>(
     key: TKey
   ): TStorage[TKey] | null {
-    return deprecated_useMap(key) as unknown as TStorage[TKey];
+    return useMap_deprecated(key) as unknown as TStorage[TKey];
   }
 
   function useObject<TKey extends Extract<keyof TStorage, string>>(
     key: TKey
   ): TStorage[TKey] | null {
-    return deprecated_useObject(key) as unknown as TStorage[TKey];
+    return useObject_deprecated(key) as unknown as TStorage[TKey];
   }
 
   function useHistory(): History {
@@ -763,6 +782,10 @@ Please see https://bit.ly/3Niy5aP for details.`
     const [root] = useStorage();
     const rerender = useRerender();
 
+    // Note: We'll hold on to the initial value given here, and ignore any
+    // changes to this argument in subsequent renders
+    const frozenInitialValue = useInitial(initialValue);
+
     React.useEffect(() => {
       if (root == null) {
         return;
@@ -771,7 +794,7 @@ Please see https://bit.ly/3Niy5aP for details.`
       let liveValue: T | undefined = root.get(key) as T | undefined;
 
       if (liveValue == null) {
-        liveValue = initialValue;
+        liveValue = frozenInitialValue;
         root.set(key, liveValue as unknown as TStorage[string]);
         //                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ FIXME
       }
@@ -804,7 +827,7 @@ Please see https://bit.ly/3Niy5aP for details.`
         unsubscribeRoot();
         unsubscribeCrdt();
       };
-    }, [root, room]);
+    }, [root, room, key, frozenInitialValue, rerender]);
 
     if (root == null) {
       return { status: "loading" };
@@ -837,8 +860,8 @@ Please see https://bit.ly/3Niy5aP for details.`
     useUndo,
     useUpdateMyPresence,
 
-    deprecated_useList,
-    deprecated_useMap,
-    deprecated_useObject,
+    useList_deprecated,
+    useMap_deprecated,
+    useObject_deprecated,
   };
 }